Added a stamina bar, and the ability to run
the blue bar at the bottom left next to health represents stamina, now,
pressing and holding shift will make the player run faster. as of now,
sprint is unlimited, but eventually, the blue bar willl decrease as you
"run" and it will slowly fill over time when not in use.