Collision
The Collision between Enemies now works. The Enemies will stop moving
once they get within an area of you. You are also pushed back whenever
they run into you. The physics seem a little off to me and I need to
debug why when the Enemy and Terrain count goes up, the speed of the
player goes up crazy. But I think this is enough for the people who need
it to work on other stuff. I have the terrain code working kinda. I
think how the terrain is created needs to be done a little differently
so that the values are stored and not randomly generated and then the
position of that terrain goes away. I thinkthat is the problem. I
commented out the collision for the terrain.

Also, I have code so that the Enemies will collide with each other, but
I personally think it's a little cooler when the enemies stack up on
each other and start massing towards you. Easy change if you want.